backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Liam Tapper",email = "97718694+tliam1@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "fastapi (>=0.119.1,<0.120.0)",
    "openai (>=2.6.0,<3.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "langchain (>=1.0.2,<2.0.0)",
    "pypdf (>=6.1.3,<7.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
]

[tool.poetry]
packages = [{ include = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "uvicorn[standard] (>=0.38.0,<0.39.0)"
]
